% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Bam.R
\name{coord_to_grng}
\alias{coord_to_grng}
\title{Convert coordinate string to grng object}
\usage{
coord_to_grng(coord)
}
\arguments{
\item{coord}{string in the form of "chr1:123-456" or "chr1:1,234-5,678,912"}
}
\value{
GRanges object corresponding to the input coordinate string
}
\description{
Convert coordinate string to grng object
}
\examples{
coord_to_grng("chr1:123-456")
coord_to_grng("chr1:1,234-5,678,912")

}
