% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_Bam.R
\name{coord_to_lst}
\alias{coord_to_lst}
\title{Convert coordinate string to list of chr,start and end}
\usage{
coord_to_lst(coord)
}
\arguments{
\item{coord}{string in the form of "chr1:123-456" or "chr1:1,234-5,678,912"}
}
\value{
a list of 3 elements. Chromosome name, start position and end position.
}
\description{
Convert coordinate string to list of chr,start and end
}
\examples{
coord_to_lst("chr1:123-456")
coord_to_lst("chr1:1,234-5,678,912")

}
