% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELViS_toy_run_result}
\alias{ELViS_toy_run_result}
\title{ELViS Toy Example - Run Result}
\format{
\subsection{\code{ELViS_toy_run_result}}{

A list of
\describe{
\item{is_reduced_output}{Indicates whether this results is a reduced form}
\item{final_output}{Final base-resolution segmentation output}
\item{final_call}{Indices of samples in which copy number variants were detected}
\item{new_Y_p2}{Normalized read depth}
}
}
}
\usage{
data(ELViS_toy_run_result)
}
\description{
List containing ELViS run result
}
\keyword{datasets}
