% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtrx_samtools_reticulate}
\alias{mtrx_samtools_reticulate}
\title{ELViS Toy Example - Base-Resolution Raw Read Depth}
\format{
\subsection{\code{mtrx_samtools_reticulate}}{

A matrix with 7906 rows and 120 columns
}
}
\usage{
data(mtrx_samtools_reticulate)
}
\description{
Base-resolution raw read depth profile over viral genome
}
\keyword{datasets}
