% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{norm.fun}
\alias{norm.fun}
\title{Normalization - scaling by certain quantile}
\usage{
norm.fun(x, probs = 0.75)
}
\arguments{
\item{x}{numeric vector to normalize.}

\item{probs}{a single numeric value of probabilities in \link{0,1} used for normalization.(Default = 0.75)}
}
\value{
numeric vector of normalized values
}
\description{
Normalization - scaling by certain quantile
}
\examples{

norm.fun(1:5)
# [1] 0.25 0.50 0.75 1.00 1.25

}
