% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{plot_pileUp}
\alias{plot_pileUp}
\title{Get a pile-up plot}
\usage{
plot_pileUp(
  x,
  target_cn_table = NULL,
  baseline = 1,
  col_cn_baseline = "#708C98",
  col_pal_cn = piratepal(palette = "info2")[-5],
  scale_plot_yaxis = TRUE
)
}
\arguments{
\item{x}{target pile-up vector to plot}

\item{target_cn_table}{final copy number table(default : NULL)}

\item{baseline}{the copy number state index you want to set as baseline}

\item{col_cn_baseline}{color for baseline}

\item{col_pal_cn}{color palette for non-baseline copy number states}

\item{scale_plot_yaxis}{Determine whether to scale the plot so that the lower bound of the y-axis is set to the lesser value between 0 and the minimum data value. (Default : TRUE)}
}
\value{
ggplot object of pile-up plot
}
\description{
Get a pile-up plot
}
\examples{

# example usage is given in the vignette (to reduce package size)

}
