% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_ELViS}
\alias{run_ELViS}
\title{Run ELViS using input raw depth matrix}
\usage{
run_ELViS(
  X,
  N_cores = min(10L, parallel::detectCores()),
  reduced_output = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{X}{Raw depth matrix of position x samples}

\item{N_cores}{THe number of cores to use (Default : \code{min(10L,parallel::detectCores())})}

\item{reduced_output}{logical indicating whether to return only reduced output}

\item{verbose}{logical indicating whether to print out messages for debugging}
}
\value{
list containing ELViS run results
}
\description{
Run ELViS using input raw depth matrix
}
\examples{

# its usage example is given in vignette in detail
}
