% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{total_aligned_base__host_and_virus}
\alias{total_aligned_base__host_and_virus}
\title{ELViS Toy Example - Total Aligned Base}
\format{
\subsection{\code{total_aligned_base__host_and_virus}}{

A vector of length 120
}
}
\usage{
data(total_aligned_base__host_and_virus)
}
\description{
Total aligned base both to host and viral genome.
}
\keyword{datasets}
