% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_example}
\alias{toy_example}
\title{ELViS Toy Example - Metadata}
\format{
\subsection{\code{toy_example}}{

A data frame with 120 rows and 6 columns:
\describe{
\item{VarType}{Variant type. Set to control if there is no variant}
\item{Copies_Altered}{The number of copies duplicated or deleted}
\item{Event_Size}{Variant size}
\item{Mean_Depth}{Mean read depth}
}
}
}
\usage{
data(toy_example)
}
\description{
Metadata of samples in the toy examples
}
\keyword{datasets}
