% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderImage.R
\name{try_renderImage}
\alias{try_renderImage}
\title{try_renderImage}
\usage{
try_renderImage(
  expr,
  env = parent.frame(),
  deleteFile = FALSE,
  outputArgs = list(),
  errorHandler = elvis::elvis_error_handler,
  warningHandler = warning
)
}
\arguments{
\item{expr}{An expression that returns a list.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{deleteFile}{Should the file in \code{func()$src} be deleted after
it is sent to the client browser? Generally speaking, if the image is a
temp file generated within \code{func}, then this should be \code{TRUE};
if the image is not a temp file, this should be \code{FALSE}. (For backward
compatibility reasons, if this argument is missing, a warning will be
emitted, and if the file is in the temp directory it will be deleted. In
the future, this warning will become an error.)}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:imageOutput]{imageOutput()}} when \code{renderImage} is used in an
interactive R Markdown document.}

\item{errorHandler, warningHandler}{passed to tryCatch}
}
\description{
try_renderImage
}
