% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderPrint.R
\name{try_renderPrint}
\alias{try_renderPrint}
\title{try_renderPrint}
\usage{
try_renderPrint(
  expr,
  env = parent.frame(),
  width = getOption("width"),
  outputArgs = list(),
  errorHandler = elvis::elvis_error_handler,
  warningHandler = warning
)
}
\arguments{
\item{expr}{An expression to evaluate.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{width}{Width of printed output.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:verbatimTextOutput]{verbatimTextOutput()}} or \code{\link[shiny:textOutput]{textOutput()}} when the functions are
used in an interactive RMarkdown document.}

\item{errorHandler, warningHandler}{passed to tryCatch}
}
\description{
try_renderPrint
}
