% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderUI.R
\name{try_renderUI}
\alias{try_renderUI}
\title{try_renderUI}
\usage{
try_renderUI(
  expr,
  env = parent.frame(),
  outputArgs = list(),
  errorHandler = elvis::elvis_error_handler,
  warningHandler = warning
)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:HTML]{HTML()}},
or a list of such objects.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:uiOutput]{uiOutput()}} when \code{renderUI} is used in an
interactive R Markdown document.}

\item{errorHandler, warningHandler}{passed to tryCatch}
}
\description{
try_renderUI
}
