% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{gene_cn_heatmaps}
\alias{gene_cn_heatmaps}
\title{Gene Copy Number Heatmap}
\usage{
gene_cn_heatmaps(
  X_raw,
  result,
  gff3_fn,
  gene_ref,
  baseline = 1,
  exclude_genes,
  col_cn = colorRamp2(c(0.5, 1, 1.5), c(muted("blue"), "white", muted("red"))),
  heatmap_height = unit(1.5, "in")
)
}
\arguments{
\item{X_raw}{Raw depth matrix}

\item{result}{Run result}

\item{gff3_fn}{gene annotation file name}

\item{gene_ref}{The name of the gene to set as reference for relative gene dosage heatmap}

\item{baseline}{Vector of state numbers to use as baseline for each sample. If it is single integer, then the given state number is used for all samples. (Default : \code{1})}

\item{exclude_genes}{name of genes to exclude from the annotation track (Default : NULL)}

\item{col_cn}{relative gene dosage color palette. (Default : \code{colorRamp2(c(0.5,1,1.5),c(muted("blue"),"white",muted("red")))})}

\item{heatmap_height}{heatmap height specified using unit function. (Default : \code{unit(1.5,"in")})}
}
\value{
a ComplexHeatmap Heatmap List object
}
\description{
Gene Copy Number Heatmap
}
\examples{

# example usage is given in the vignette (to reduce package size)

}
