% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_ELViS}
\alias{run_ELViS}
\title{Run ELViS using input raw depth matrix}
\usage{
run_ELViS(
  X,
  N_cores = min(10L, parallel::detectCores()),
  reduced_output = TRUE
)
}
\arguments{
\item{X}{Raw depth matrix of position x samples}

\item{N_cores}{THe number of cores to use (Default : \code{min(10L,parallel::detectCores())})}

\item{reduced_output}{logical indicating whether to return only reduced output}
}
\value{
list containing ELViS run results
}
\description{
Run ELViS using input raw depth matrix
}
\examples{

data(mtrx_samtools_reticulate)
th<-50
filtered_mtrx <- filt_samples(mtrx_samtools_reticulate,th=th,smry_fun=max)

result <- run_ELViS(filtered_mtrx[,seq_len(5)],N_cores=1L)

}
