% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderDataTable.R
\name{try_renderDataTable}
\alias{try_renderDataTable}
\title{try_renderDataTable}
\usage{
try_renderDataTable(
  expr,
  options = NULL,
  searchDelay = 500,
  callback = "function(oTable) {}",
  escape = TRUE,
  env = parent.frame(),
  outputArgs = list(),
  errorHandler = elvis::elvis_error_handler,
  warningHandler = warning
)
}
\arguments{
\item{expr}{An expression that generates a plot.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link[shiny:plotOutput]{plotOutput()}} when \code{renderPlot} is used in an
interactive R Markdown document.}

\item{errorHandler, warningHandler}{passed to tryCatch}
}
\description{
try_renderDataTable
}
